% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{calcCorr}
\alias{calcCorr}
\title{Check model design and calculate correlation matrix. This is an internal function.  Use runCorr() instead (which is a wrapper to this function).}
\usage{
calcCorr(modeldata, metabdata, cohort = "")
}
\arguments{
\item{modeldata}{list from function getModelData}

\item{metabdata}{metabolite data list}

\item{cohort}{cohort label (e.g DPP, NCI, Shanghai)}
}
\value{
data frame with each row representing the correlation for each combination of outcomes and exposures represented as specified in the
model (*spec), label (*lab), and universal id (*_uid)
with additional columns for n, pvalue, method of model specification (Interactive or Batch), universal id for outcomes () and exposures
name of the cohort and adjustment variables. Attribute of dataframe includes ptime for processing time of model
run.
}
\description{
Check model design and calculate correlation matrix. This is an internal function.  Use runCorr() instead (which is a wrapper to this function).
}
