% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{checkModelDesign}
\alias{checkModelDesign}
\title{Checks integrity of sheets in the user input CSV file}
\usage{
checkModelDesign(modeldata = NULL, createDummies = NULL)
}
\arguments{
\item{modeldata}{(output of function getModelData())}

\item{crateDummies}{TRUE or FALSE}
}
\value{
modeldata after checks are performed
}
\description{
Checks integrity of sheets in the user input CSV file
}
\keyword{internal}
