% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{CometsAnalytics}
\alias{CometsAnalytics}
\alias{CometsAnalytics-package}
\title{COMETS Analytics R package}
\description{
This R package support all cohort-specific analyses of
       the COMETS consortium. Data are not saved in the system but
       output must be downloaded and submitted for meta-analyses.
 import only functions needed
}
