% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_createCOMETSinput.R
\name{createCOMETSinput}
\alias{createCOMETSinput}
\title{Read in CSV files and create an Excell file that's formatted for input into Comets Analytics
Once the output Excell file is created, users will need to go in and complete
the sheets 'VarMap' and 'Models'}
\usage{
createCOMETSinput(metabfile = NULL, subjfile = NULL,
  abundancesfile = NULL, outputfile = NULL)
}
\arguments{
\item{metabfile}{CSV file name, including path, of metabolite meta data}

\item{subjfile}{CSV file name, includingpath, of subject meta data}

\item{abundancesfile}{CSV file name, including path, of metabolite abundances}

\item{outputfile}{name, including path,  of output .xlsx file}
}
\description{
Read in CSV files and create an Excell file that's formatted for input into Comets Analytics
Once the output Excell file is created, users will need to go in and complete
the sheets 'VarMap' and 'Models'
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
metabfile <- file.path(dir, "testmetab.csv")
subjfile <- file.path(dir, "testsubject.csv")
abundancesfile <- file.path(dir, "testabundances.csv")
createCOMETSinput(metabfile=metabfile, 
      subjfile=subjfile,
      abundancesfile=abundancesfile,
      outputfile="MyData.xlsx")

}

