% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_filterData.R
\name{filterCOMETSinput}
\alias{filterCOMETSinput}
\title{Optional function that subsets input data (e.g. only perform analysis on female subjects)}
\usage{
filterCOMETSinput(readData, where = NULL)
}
\arguments{
\item{readData}{list from readComets}

\item{where}{users can specify which subjects to perform the analysis by specifying this parameter. 'where' expects a vector with a variable name, a comparison operator (e.g. "<", ">", "="), and a value.  For example, "where = c("Gender","=","Female")}
}
\value{
filtered list
}
\description{
Optional function that subsets input data (e.g. only perform analysis on female subjects)
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
filtexmetabdata <- filterCOMETSinput(exmetabdata, where <- c("age",">","60"))

}
