% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{fixData}
\alias{fixData}
\title{Fixes input CSV data (e.g. takes care of factors, and other data frame conversions)}
\usage{
fixData(dta, compbl = FALSE)
}
\arguments{
\item{dta}{any data frame}

\item{compbl}{compress multiple blank spaces to single blank space for all character or factor variables in the dataset}
}
\description{
Fixes input CSV data (e.g. takes care of factors, and other data frame conversions)
}
\keyword{internal}
