% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.r
\name{getModelData}
\alias{getModelData}
\title{Get data ready for the model runs in interactive or batch mode as specified in the input}
\usage{
getModelData(readData, modelspec = "Batch", modbatch = "",
  rowvars = "All metabolites", colvars = "", adjvars = NULL)
}
\arguments{
\item{readData}{list from readComets}

\item{modelspec}{How model is specified (Interactive or Batch)}

\item{modbatch}{if batch, chosen model specified by batch mode}

\item{rowvars}{if Interactive, outcome variables (usually metabolites and rendered in rows default in All metabolites)}

\item{colvars}{if Interactive, exposure variables (usually covariates rendered in columns)}

\item{adjvars}{If Interactive, adjustment covariates}
}
\value{
a list comprising:

1: subset data: gdta

2: column variables: ccovs

3: row variables: rcovs

4: adjustment variables: acovs
}
\description{
Get data ready for the model runs in interactive or batch mode as specified in the input
}
\examples{

modeldata <- getModelData(readData(),modelspec="Batch",colvars="age")

}

