% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{HarmonizeNew}
\alias{HarmonizeNew}
\title{Fixes input CSV data (e.g. takes care of factors, and other data frame conversions)}
\usage{
HarmonizeNew(dtalist, mastermetid)
}
\arguments{
\item{dtalist}{results of reading a CSV data sheet (with read_excel)}

\item{matermetid}{processed uid.csv file}
}
\description{
Fixes input CSV data (e.g. takes care of factors, and other data frame conversions)
}

