% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOutput.r
\name{makeOutputCSV}
\alias{makeOutputCSV}
\title{Create output CSV file}
\usage{
makeOutputCSV(file, dataf, cohort = "")
}
\arguments{
\item{file}{filename}

\item{dataf}{dataframe}

\item{cohort}{cohort name}
}
\value{
a CSV file with results named with cohort
}
\description{
Create output CSV file
}
\examples{
dir <- system.file("extdata", package="CometsAnalyticsPackage", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
modeldata <- getModelData(readCSV(csvfile))
corrmatrix <-getCorr(modeldata)
}

