% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOutput.r
\name{OutputCSVResults}
\alias{OutputCSVResults}
\title{Create output CSV file}
\usage{
OutputCSVResults(filename, dataf, cohort = "")
}
\arguments{
\item{filename}{if type_output is "CSVfile", name of file}

\item{dataf}{correlation output (from function getCorr())}

\item{cohort}{cohort name}
}
\value{
the filename of the CSV file with results named with cohort
}
\description{
Create output CSV file
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modbatch="1.1 Unadjusted")
corrmatrix <-getCorr(modeldata,exmetabdata,"DPP")
OutputCSVResults(filename="corr",dataf=corrmatrix,cohort="DPP") #get correlation results
OutputCSVResults(filename="harmonization",dataf=exmetabdata$metab,cohort="DPP") #get harmonization results
}

