% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingfunctions.r
\name{OutputXLSResults}
\alias{OutputXLSResults}
\title{Create output XLSX file}
\usage{
OutputXLSResults(filename, datal, cohort = "")
}
\arguments{
\item{filename}{name of file and can include path}

\item{datal}{data list to output (each item on list outputs to a worksheet)}

\item{cohort}{cohort name}
}
\value{
the filename of the XLSX file with results named with cohort
}
\description{
Create output XLSX file
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAgeTest.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modlabel="1 Gender adjusted")
# Get descriptive data
descdata <-runDescrip(exmetabdata)
OutputXLSResults(filename="corr",datal=descdata,cohort="DPP")
}
