% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotVar}
\alias{plotVar}
\title{Plot the variance distribution of transformed metabolite abundances}
\usage{
plotVar(cometsdata, title = "Distribution of Variance for Each Metabolite",
  xlabel = "Variance of transformed metabolite abundances",
  ylabel = "Frequency")
}
\arguments{
\item{cometsdata}{output of readCOMETSinput function}

\item{title}{main title for the plot (default is "Distribution of Variance for Each Metabolite")}

\item{xlabel}{x-axis label (default is "Variance of transformed metabolite abundances")}

\item{ylabel}{y-axis label (default is "Frequency")}
}
\value{
a distribution plot
}
\description{
Plot the variance distribution of transformed metabolite abundances
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
plotVar(exmetabdata)

}

