% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_readCOMETSinput.R
\name{readCOMETSinput}
\alias{readCOMETSinput}
\title{Read in Excell file that contains metabolite abundances and metab data}
\usage{
readCOMETSinput(csvfilePath)
}
\arguments{
\item{csvfilePath}{path of Excell file to be read in}
}
\value{
a list comprising:
}
\description{
Read in Excell file that contains metabolite abundances and metab data
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)

}
