% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCSV.r
\name{readCSV}
\alias{readCSV}
\title{Read in CSV file that contains metabolite abundances and metab data}
\usage{
readCSV(csvfilePath, modelspec = "NoBATCH")
}
\arguments{
\item{csvfilePath}{path of CSV file to be read in}

\item{modelspec}{if "BATCH", then runs multiple models (default is "NoBATCH")}
}
\value{
a list comprising:
}
\description{
Read in CSV file that contains metabolite abundances and metab data
}
\examples{
dir <- system.file("extdata", package="CometsAnalyticsPackage", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
metabdata <- readCSV(csvfile)

}

