% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_RunAllModels.R
\name{runAllModels}
\alias{runAllModels}
\title{This function allows users to run all models that are provided in the "Models" sheet of
the input Excell file.}
\usage{
runAllModels(readData, cohort = "", writeTofile = T)
}
\arguments{
\item{readData}{list from readComets}

\item{cohort}{cohort label (e.g. DPP, NCI, Shanghai)}

\item{writeTofile}{T/F (whether or not to write results for each model onto
separate xls file). Files are written to current directory. Default is True.}
}
\value{
a list of data frames, where each data frame has rows representing the correlation for each combination of outcomes and exposures with additional columns for n, pvalue, metabolite_id, method of model specification (Interactive or Batch), name of the cohort and adjustment variables. Attribute of dataframe includes ptime for processing time of model run.
}
\description{
This function allows users to run all models that are provided in the "Models" sheet of
the input Excell file.
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
allmodeloutput <- runAllModels(exmetabdata)

}
