% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getCorr.r
\name{runCorr}
\alias{runCorr}
\title{Calculate correlation matrix for input model. This function will check for strata and if present, it will run the correlation matrix within each strata. Each model design is checked for validity (correlation between predictors, zero variance, etc.).}
\usage{
runCorr(modeldata, metabdata, cohort = "")
}
\arguments{
\item{modeldata}{list from function getModelData}

\item{metabdata}{metabolite data list}

\item{cohort}{cohort label (e.g DPP, NCI, Shanghai)}
}
\value{
data frame with each row representing the correlation for each combination of outcomes and exposures represented as specified in the
model (*spec), label (*lab), and universal id (*_uid)
with additional columns for n, pvalue, method of model specification (Interactive or Batch), universal id for outcomes (outcome_uid) and exposures (exposure_uid)
name of the cohort, adjustment (adjvars) and stratification (stratavar,strata)  variables. Attribute of dataframe includes ptime for processing time of model
run.
}
\description{
Calculate correlation matrix for input model. This function will check for strata and if present, it will run the correlation matrix within each strata. Each model design is checked for validity (correlation between predictors, zero variance, etc.).
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modlabel="1 Gender adjusted",
	rowvars=c("lactose","lactate"))
corrmatrix <- runCorr(modeldata,exmetabdata, "DPP")
}
