% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_RunAllModels.R
\name{runDescrip}
\alias{runDescrip}
\title{This function allows users to run all models that are provided in the "Models" sheet of
the input Excell file.}
\usage{
runDescrip(readData)
}
\arguments{
\item{readData}{list from readComets}
}
\value{
a list with 2 data frames, continuous and categorical summaries. Type of variable is defined in varmap
}
\description{
This function allows users to run all models that are provided in the "Models" sheet of
the input Excell file.
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
allmodeloutput <- runAllModels(exmetabdata)
# Get descriptive data
descdata <-runDescrip(exmetabdata)
OutputXLSResults(filename="corr",datal=descdata,cohort="DPP")

}
