% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.r
\name{showHClust}
\alias{showHClust}
\title{Show interactive heatmap using d3heatmap with hierarchical clustering}
\usage{
showHClust(ccorrmat, clust = TRUE, colscale = "RdYlBu")
}
\arguments{
\item{ccorrmat}{correlation matrix}

\item{clust}{Show hierarchical clustering}

\item{colscale}{colorscale, can be custom or named ("Hots","Greens","Blues","Greys","Purples") see \url{https://plot.ly/ipython-notebooks/color-scales/}}
}
\value{
a heatmap with outcomes as rows and exposures in columns.
}
\description{
Show interactive heatmap using d3heatmap with hierarchical clustering
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,modbatch="1.1 Unadjusted")
corrmatrix <-getCorr(modeldata,exmetabdata,"DPP")
showHClust(corrmatrix)
}
\references{
For colorscale reference: \url{https://plot.ly/ipython-notebooks/color-scales/}
}

