% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getCorr.r
\name{stratCorr}
\alias{stratCorr}
\title{Calculate correlation matrix for each strata specified if stratification is specified in the model tab or in interactive mode}
\usage{
stratCorr(modeldata, metabdata, cohort = "")
}
\arguments{
\item{modeldata}{list from function getModelData}

\item{metabdata}{metabolite data list}

\item{cohort}{cohort label (e.g DPP, NCI, Shanghai)}
}
\value{
data frame with each row representing the correlation for each combination of outcomes and exposures represented as specified in the
model (*spec), label (*lab), and universal id (*_uid)
with additional columns for n, pvalue, method of model specification (Interactive or Batch), universal id for outcomes (outcome_uid) and exposures (exposure_uid)
name of the cohort, adjustment (adjvars) and stratification (stratavar,strata)  variables. Attribute of dataframe includes ptime for processing time of model
run.
}
\description{
Calculate correlation matrix for each strata specified if stratification is specified in the model tab or in interactive mode
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modbatch="1.1 Unadjusted")
corrmatrix <-getCorr(modeldata,exmetabdata, "DPP")
}
