% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{checkModelDesign}
\alias{checkModelDesign}
\title{Ensures that models will run without errors.  Preprocesses design matrix for 
zero variance, linear combinations, and dummies.}
\usage{
checkModelDesign(modeldata = NULL)
}
\arguments{
\item{modeldata}{(output of function getModelData())}
}
\value{
modeldata after checks are performed
}
\description{
Ensures that models will run without errors.  Preprocesses design matrix for 
zero variance, linear combinations, and dummies.
}
\keyword{internal}
