% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_createCOMETSinput.R
\name{createCOMETSinput}
\alias{createCOMETSinput}
\title{Read in CSV files and create an Excell file that's formatted for input into Comets Analytics
Once the output Excell file is created, users will need to go in and complete
the sheets 'VarMap' and 'Models'}
\usage{
createCOMETSinput(filenames = NULL, outputfile = NULL)
}
\arguments{
\item{filenames}{names of CSV files, including path, of metabolite meta data,
metabolite abundances, and subject meta data (in that order)}

\item{outputfile}{name, including path,  of output .xlsx file}
}
\description{
Read in CSV files and create an Excell file that's formatted for input into Comets Analytics
Once the output Excell file is created, users will need to go in and complete
the sheets 'VarMap' and 'Models'
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
metabfile <- file.path(dir, "testmetab.csv")
subjfile <- file.path(dir, "testsubject.csv")
abundancesfile <- file.path(dir, "testabundances.csv")
createCOMETSinput(filenames=c(metabfile,abundancesfile,subjfile), 
      outputfile="MyData.xlsx")

}

