% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{filterCOMETSinput}
\alias{filterCOMETSinput}
\title{Function that subsets input data based on "where variable"}
\usage{
filterCOMETSinput(readData, where = NULL)
}
\arguments{
\item{readData}{list from readComets}

\item{where}{users can specify which subjects to perform the analysis by specifying this parameter. 'where' expects a vector with a variable name, a comparison operator (e.g. "<", ">", "="), and a value.  For example, "where = c("Gender","=","Female")}
}
\value{
filtered list
}
\description{
Function that subsets input data based on "where variable"
}
