% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{fixData}
\alias{fixData}
\title{Fixes input CSV data (e.g. takes care of factors, and other data frame conversions)}
\usage{
fixData(dta)
}
\arguments{
\item{dta}{results of reading a CSV data sheet (with read_excel)}
}
\description{
Fixes input CSV data (e.g. takes care of factors, and other data frame conversions)
}
\keyword{internal}
