% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.r
\name{getModelData}
\alias{getModelData}
\title{Get data ready for the model runs in interactive or batch mode as specified in the input}
\usage{
getModelData(readData, rowvars = "All metabolites", colvars = "",
  adjvars = NULL)
}
\arguments{
\item{readData}{list from readComets}

\item{rowvars}{Outcome variables (usually metabolites and rendered in rows default in All metabolites)}

\item{colvars}{Exposure variables (usually covariates rendered in columns)}

\item{adjvars}{Adjustment covariates}
}
\value{
a list comprising:

1: subset data: gdta

2: column variables: ccovs

3: row variables: rcovs

4: adjustment variables: acovs
}
\description{
Get data ready for the model runs in interactive or batch mode as specified in the input
}
\examples{

modeldata <- getModelData(readData(),modbatch="Interactive",colvars="age")

}

