% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_getModelData.R
\name{getModelData}
\alias{getModelData}
\title{Get data ready for the model runs in interactive or batch mode as specified in the input}
\usage{
getModelData(readData, modelspec = "Batch", modlabel = "",
  rowvars = "All metabolites", colvars = "", adjvars = NULL,
  strvars = NULL, where = NULL)
}
\arguments{
\item{readData}{list from readComets}

\item{modelspec}{How model is specified (Interactive or Batch)}

\item{modlabel}{if batch, chosen model specified by batch mode. If interactive model label.}

\item{rowvars}{if Interactive, outcome variables (usually metabolites rendered in rows, default is All metabolites)}

\item{colvars}{if Interactive, exposure variables (usually covariates rendered in columns)}

\item{adjvars}{If Interactive, adjustment covariates}

\item{strvars}{If Interactive, stratification covariates}

\item{where}{users can specify which subjects to perform the analysis by specifying this parameter. 'where' expects a vector of strings with a variable name, a comparison operator (e.g. "<", ">", "="), and a value.  For example, "where = c("age>50","bmi > 22").  Note that rules must be separate by a comma.}
}
\value{
a list comprising:

1: subset data: gdta

2: column variables: ccovs

3: row variables: rcovs

4: adjustment variables: acovs

5: stratification variable: scovs

6: model specification: modspec

6: model label: modlab
}
\description{
Get data ready for the model runs in interactive or batch mode as specified in the input
}
\examples{

dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modlabel="1.1 Unadjusted")

}
