% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{Harmonize}
\alias{Harmonize}
\title{Harmonizes metabolites by looking up metabolites names from user input and finding the corresponding COMETS harmonized name.}
\usage{
Harmonize(dtalist)
}
\arguments{
\item{dtalist}{results of reading a CSV data sheet (with read_excel)}
}
\description{
Harmonizes metabolites by looking up metabolites names from user input and finding the corresponding COMETS harmonized name.
}
\keyword{internal}
