% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOutput.r
\name{makeOutputCSV}
\alias{makeOutputCSV}
\title{Create output CSV file}
\usage{
makeOutputCSV(type_output = "display", filename, dataf, cohort = "")
}
\arguments{
\item{type_output}{"display" or "CSVfile"}

\item{filename}{if type_output is "CSVfile", name of file}

\item{dataf}{correlation output (from function getCorr()) data frame}

\item{cohort}{cohort name}
}
\value{
the filename of the CSV file with results named with cohort
}
\description{
Create output CSV file
}
\examples{
dir <- system.file("extdata", package="CometsAnalyticsPackage", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modbatch="1.1 Unadjusted")
corrmatrix <-getCorr(modeldata,exmetabdata,"DPP")
makeOutputCSV(type_output="CSVfile",filename="corr",dataf=corrmatrix,cohort="DPP")
}

