% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_readCOMETSinput.R
\name{readCOMETSinput}
\alias{readCOMETSinput}
\title{Read in CSV file that contains metabolite abundances and metab data}
\usage{
readCOMETSinput(csvfilePath, modelspec = "Interactive")
}
\arguments{
\item{csvfilePath}{path of CSV file to be read in}

\item{modelspec}{if "BATCH", then runs multiple models (default is "NoBATCH")}
}
\value{
a list comprising:
}
\description{
Read in CSV file that contains metabolite abundances and metab data
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
exmetabdata <- readCOMETSinput(csvfile)

}
