% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.r
\name{showHeatmap}
\alias{showHeatmap}
\title{Show interactive heatmap using plot_ly}
\usage{
showHeatmap(ccorrmat, rowsortby = "corr", plothgt = 700, plotwid = 800,
  colscale = "RdYlBu")
}
\arguments{
\item{ccorrmat}{correlation matrix (output of getCorr())}

\item{rowsortby}{How row labels are sorted}

<<<<<<< HEAD
\item{plothgt}{Plot height default 700}

\item{plotwid}{Plot width}

\item{colscale}{colorscale, can be custom or named ("Hots","Greens","Blues","Greys","Purples") see \url{https://plot.ly/ipython-notebooks/color-scales/}}
=======
\item{plothgt}{height of plot for display (default: 700)}
>>>>>>> 39b72e7d59b78e7abdccd4f339947b902d82c05b
}
\value{
a heatmap with outcomes as rows and exposures in columns.
}
\description{
Show interactive heatmap using plot_ly
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,modbatch="1.1 Unadjusted")
corrmatrix <-getCorr(modeldata,exmetabdata,"DPP")
showHeatmap(corrmatrix)
}
\references{
For colorscale reference: \url{https://plot.ly/ipython-notebooks/color-scales/}
}

