% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.r
\name{showHeatmap}
\alias{showHeatmap}
\title{Show interactive heatmap using plot_ly}
\usage{
showHeatmap(ccorrmat, rowsortby = "metasc", plothgt = 700)
}
\arguments{
\item{ccorrmat}{correlation matrix}

\item{rowsortby}{How row labels are sorted}
}
\value{
a heatmap with outcomes as rows and exposures in columns.
}
\description{
Show interactive heatmap using plot_ly
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInput.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modbatch="1.1 Unadjusted")
corrmatrix <-getCorr(modeldata,exmetabdata,"DPP")
showHeatmap(corrmatrix)
}

