% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{showHeatmap}
\alias{showHeatmap}
\title{Show interactive heatmap using plot_ly}
\usage{
showHeatmap(ccorrmat, rowsortby = "corr", plothgt = 700, plotwid = 800,
  colscale = "RdYlBu")
}
\arguments{
\item{ccorrmat}{correlation matrix (output of runCorr())}

\item{rowsortby}{How row labels are sorted}

\item{plothgt}{Plot height default 700}

\item{plotwid}{Plot width default 800}

\item{colscale}{colorscale, can be custom or named ("Hots","Greens","Blues","Greys","Purples") see \url{https://plot.ly/ipython-notebooks/color-scales/}}
}
\value{
a heatmap with outcomes as rows and exposures in columns.
}
\description{
Show interactive heatmap using plot_ly
}
\examples{
\dontrun{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAgeTest.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,modlabel="1 Gender adjusted")
corrmatrix <-runCorr(modeldata,exmetabdata,"DPP")
showHeatmap(corrmatrix)
}
}
\references{
For colorscale reference: \url{https://plot.ly/ipython-notebooks/color-scales/}
}
